// src/user/user.dto.ts
export interface IUserProfileDto {

  first_name: string;
  last_name: string;
  department: string;
  designation: string;
  image_url: string;
  city: string;
  country: string;
  bio: string;
  social_links: string;
  employee_id: string;
}

export interface IUserUpdateDto {
  first_name: string;
  last_name: string;
  department: string;
  designation: string;
  image_url: string;
  city: string;
  country: string;
  bio: string;
  social_links: string;
 
}